tests: update eip7742 tests.
